Convert Private to Typescript functional component
